resolve challenge 1/ fix next/previos btns
