+++ pages/index.js
@@ -2,7 +2,7 @@ import Head from "next/head";
 import Image from "next/image";
 import styles from "../styles/Home.module.css";
 import Profile from "../react_challenges/describing_ui/Profile";
-import Gallery from "../react_challenges/adding_interactivity/state_comp_memory/Gallery";
+//import Gallery from "../react_challenges/adding_interactivity/state_comp_memory/Gallery";
 import List from "../react_challenges_Paula/describing-the-UI/rendering-lists/List";
 import RecipeList from "../react_challenges_Paula/describing-the-UI/rendering-lists/RecipeList";
 import Poem from "../react_challenges_Paula/describing-the-UI/rendering-lists/Poem";
@@ -10,6 +10,8 @@ import Clock from "../react_challenges_Paula/describing-the-UI/keeping-comp-pure
 import ProfileList from "../react_challenges_Paula/describing-the-UI/keeping-comp-pure/ProfileList";
 import CollapseDiv from "../react_challenges_Paula/describing-the-UI/keeping-comp-pure/CollapseDiv";
 import StoryTray from "../react_challenges_Paula/describing-the-UI/keeping-comp-pure/StoryTray";
+import pGallery from "../react_challenges_Paula/describing-the-UI/state-a-comp-memory/Pgallery";
+import Pgallery from "../react_challenges_Paula/describing-the-UI/state-a-comp-memory/Pgallery";
 
 export default function Home() {
   return (
@@ -32,14 +34,15 @@ export default function Home() {
         <RecipeList /> 
         <Clock /> 
         <ProfileList />*/}
-        <Clock />
-        <StoryTray
+        {/* <Clock /> */}
+        {/* <StoryTray
           stories={[
             { id: "Mike", label: "story" },
             { id: "Daniel", label: "story" },
             { id: "Daiana", label: "story" },
           ]}
-        />
+        /> */}
+        <Pgallery />
       </main>
 
       <footer className={styles.footer}>
