+++ styles/globals.css
@@ -67,3 +67,6 @@ a {
   height: 100px;
 }
 
+.hidden{
+  visibility: hidden;
+}
